# docs/swagger.yaml
basePath: /api
definitions:
  domain.Author:
    properties:
      avatar_emoji:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  domain.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/domain.Author'
      body:
        type: string
      created_at:
        type: string
      id:
        type: integer
    type: object
  domain.CreateCommentRequest:
    properties:
      body:
        example: Great analysis!
        type: string
    type: object
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.PaginatedResponse:
    properties:
      data: {}
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for Trading Chat application with user authentication and
    management.
  title: Trading Chat API
  version: "1.0"
paths:
  /comments/{id}:
    delete:
      description: Delete a comment (only owner can delete)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - Comments
...
  /posts/{id}/comments:
    get:
      description: Get all comments for a post, paginated
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated comments
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get comments for a post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a new comment on a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created comment
          schema:
            $ref: '#/definitions/domain.CommentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a comment to a post
      tags:
      - Comments
  /posts/{id}/like:
    delete:
      description: Unlike a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlike a post
      tags:
      - Likes
    post:
      description: Like a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Like a post
      tags:
      - Likes
...
  /users/{id}/follow:
    delete:
      description: Unfollow a user
      parameters:
      - description: User ID to unfollow
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - Followers
    post:
      description: Follow another user
      parameters:
      - description: User ID to follow
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - Followers
  /users/{id}/followers:
    get:
      description: Get list of users who follow the specified user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated followers
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's followers
      tags:
      - Followers
  /users/{id}/following:
    get:
      description: Get list of users the specified user is following
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated following
          schema:
            $ref: '#/definitions/domain.PaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's following
      tags:
      - Followers
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"